backend/requirements.txt
# Basis-Stack für FastAPI, ClickHouse, Docker & Whale-Detection

# --- Standard ---
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
fastapi==0.115.13
h11<0.15,>=0.13
httptools==0.6.4
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.0
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
clickhouse-connect==0.7.9

# --- Empfohlene Erweiterungen für Web, Requests, EVM/Krypto ---
httpx>=0.26.0        # Async HTTP-Client für REST, Webhook, API-Calls
web3==6.19.0           # EVM-Integration, Whale-Detection, Ethereum/WebSocket

# --- (Optional, für Bilder, AI, etc.) ---
pillow==10.3.0        # Images/Screenshots
# openai==1.30.1        # AI-Integration, wenn OpenAI genutzt wird

# --- (Optional: Benachrichtigungen, Scheduler, Telegram etc.) ---
python-telegram-bot==20.8  # Für Telegram-Nachrichten
aiojobs==1.2.1            # Für asynchrone Scheduled Tasks (Cron/Scheduler)
